requirements.txt
Django 
asgiref
autopep8
pycodestyle
pytz
sqlparse
Unipath
dj-database-url
python-decouple
gunicorn
whitenoise
Pillow
django-allauth
django-bootstrap-form
django-import-export
django-admin-shortcuts
django-extra-settings
django-jalali-date
django-mptt
